pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "salestrust_pipeline"
version = "0.1.0"
authors = [
  { name = "tmbb", email = "thatianembb@gmail.com" }
]
description = "Pipeline de dados em PySpark para geração de relatório de pedidos com pagamentos recusados e legítimos."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]
dependencies = [
  "pyspark==4.1.1",
  "pyyaml==6.0.3"
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.1",
  "ruff==0.12.9",
  "black==25.1.0",
  "build==1.3.0"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}